requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Cors==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Financial Data
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# Background Jobs
APScheduler==3.10.4
celery==5.3.4

# Authentication & Security
Flask-JWT-Extended==4.5.3
bcrypt==4.1.2
email-validator==2.1.0

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
python-decouple==3.8

# Charts & Visualization (for server-side rendering if needed)
plotly==5.18.0

# Email notifications
Flask-Mail==0.9.1